device: add first draft of RThermo device
This device implementation is done on Nordic Semiconductor nRF24LE1
SoC which embed a 8051 MCU and a nRF24L01+ RF chipset.

It implements basics of RThermo protocol which are pairing and
periodic status requests.
At first boot (after flash), it generates a unique random serial
number which is stored in flash memory.
A button is dedicated to pairing requests triggering from user. The
MCU is waken up and try to do a pairing  handshaking with main
controller in order to get an address. If it succeed, the address
assigned by main controller is also saved in flash memory.
When address is assigned to device, it wakes up at a periodic rate
(arround a minute) to send a status request and execute necessary
actions asked by main controller like toggling heaters.

By using low power mode ("memory retention with timer on" mode), it
intends to save as much as possible power and so works on a simple LR6
battery for more than one year.