linux: add first draft of RThermo main controller
This main controller implementation is done on Linux system with a
nRF24L01(+) connected to the machine executing Linux.

It basically uses two libraries: libconfig to load and store main
configuration and sqlite3 to store dynamic data like associated
devices and their configuration, and all temperature monitoring data.

For nRF24L01(+) control, it uses basic GPIO / IRQ sysfs entries of
Linux kernel GPIO sysfs class and SPIDEV driver of Linux kernel.
Some helpers can be found in src/sys folder. Everything specific to
the chipset is managed by the nrf24 class for easy usage.

The main controller is another class which is completly independent:
it creates its own thread to run separatly from the rest of the daemon
which can runs an HTTP server instance. It implements basics of
RThermo protocol which are pairing and status requests handling and
response generation.
The main loop can be set in pairing mode to wait for incoming pairing
requests from one or more devices, or in normal mode which listen for
up to six devices.
At first launch (no database created) a serial random and an address
are randomly generated and stored in database.

Some sample configuration file are provided for Standard Linux,
Raspberry Pi 3 and Jetson TX1 boards.